telegram_bot/requirements.txt
python-telegram-bot==13.15
asyncpg
openai
requests
python-dotenv
urllib3<2.0
